+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## lifebit-ai/cloudos-cli: changelog
 
 ### 1.3.0 - 2022-11-07
+- All Cromwell functionality works now with personal API key. The
+`--cromwell-token` argument is maintained for backwards compatibility, but can
+be completely substituted by `--apikey`.
 - Changes `--wdl-importsfile` parameter to be optional even when running a
 WDL pipeline as `importsFiles` are not always present in WDL pipelines.
 - Fixes some incomplete error messages.